+++ package.js
@@ -1,90 +1,123 @@
-/* eslint no-shadow: 0, func-names: 0, no-unused-vars: 0, no-console: 0 */
-var os = require('os');
-var webpack = require('webpack');
-var cfg = require('./webpack.config.production.js');
-var packager = require('electron-packager');
-var assign = require('object-assign');
-var del = require('del');
-var exec = require('child_process').exec;
-var argv = require('minimist')(process.argv.slice(2));
-var devDeps = Object.keys(require('./package.json').devDependencies);
+/* eslint strict: 0, no-shadow: 0, no-unused-vars: 0, no-console: 0 */
+'use strict';
+
+require('babel-polyfill');
+const os = require('os');
+const webpack = require('webpack');
+const electronCfg = require('./webpack.config.electron');
+const cfg = require('./webpack.config.production');
+const packager = require('electron-packager');
+const del = require('del');
+const exec = require('child_process').exec;
+const argv = require('minimist')(process.argv.slice(2));
+const pkg = require('./package.json');
+
+const deps = Object.keys(pkg.dependencies);
+const devDeps = Object.keys(pkg.devDependencies);
+
+const appName = argv.name || argv.n || pkg.productName;
+const shouldUseAsar = argv.asar || argv.a || false;
+const shouldBuildAll = argv.all || false;
 
 
-var appName = argv.name || argv.n || 'ElectronReact';
-var shouldUseAsar = argv.asar || argv.a || false;
-var shouldBuildAll = argv.all || false;
-
-
-var DEFAULT_OPTS = {
+const DEFAULT_OPTS = {
   dir: './',
   name: appName,
   asar: shouldUseAsar,
   ignore: [
-    '/test($|/)',
-    '/tools($|/)',
-    '/release($|/)'
-  ].concat(devDeps.map(function(name) { return '/node_modules/' + name + '($|/)'; }))
+    '^/test($|/)',
+    '^/release($|/)',
+    '^/main.development.js'
+  ].concat(devDeps.map(name => `/node_modules/${name}($|/)`))
+  .concat(
+    deps.filter(name => !electronCfg.externals.includes(name))
+      .map(name => `/node_modules/${name}($|/)`)
+  )
 };
 
-var icon = argv.icon || argv.i || 'app/app.icns';
+const icon = argv.icon || argv.i || 'app/app';
 
 if (icon) {
   DEFAULT_OPTS.icon = icon;
 }
 
-var version = argv.version || argv.v;
+const version = argv.version || argv.v;
 
 if (version) {
   DEFAULT_OPTS.version = version;
   startPack();
 } else {
   // use the same version as the currently-installed electron-prebuilt
-  exec('npm list | grep electron-prebuilt', function(err, stdout, stderr) {
+  exec('npm list electron --dev', (err, stdout) => {
     if (err) {
-      DEFAULT_OPTS.version = '0.34.0';
+      DEFAULT_OPTS.version = '1.2.0';
     } else {
-      DEFAULT_OPTS.version = stdout.split('@')[1].replace(/\s/g, '');
+      DEFAULT_OPTS.version = stdout.split('electron@')[1].replace(/\s/g, '');
     }
+
     startPack();
   });
 }
 
 
-function startPack() {
-  console.log('start pack...');
-  webpack(cfg, function runWebpackBuild(err, stats) {
-    if (err) return console.error(err);
-    del('release')
-    .then(function(paths) {
-      if (shouldBuildAll) {
-        // build for all platforms
-        var archs = ['ia32', 'x64'];
-        var platforms = ['linux', 'win32', 'darwin'];
-
-        platforms.forEach(function(plat) {
-          archs.forEach(function(arch) {
-            pack(plat, arch, log(plat, arch));
-          });
-        });
-      } else {
-        // build for current platform only
-        pack(os.platform(), os.arch(), log(os.platform(), os.arch()));
-      }
-    })
-    .catch(function(err) {
-      console.error(err);
+function build(cfg) {
+  return new Promise((resolve, reject) => {
+    webpack(cfg, (err, stats) => {
+      if (err) return reject(err);
+      resolve(stats);
     });
   });
+}
+
+async function startPack() {
+  console.log('start pack...');
+
+  try {
+    await build(electronCfg);
+    await build(cfg);
+    const paths = await del('release');
+
+    if (shouldBuildAll) {
+      // build for all platforms
+      const archs = ['ia32', 'x64'];
+      const platforms = ['linux', 'win32', 'darwin'];
+
+      platforms.forEach((plat) => {
+        archs.forEach((arch) => {
+          pack(plat, arch, log(plat, arch));
+        });
+      });
+    } else {
+      // build for current platform only
+      pack(os.platform(), os.arch(), log(os.platform(), os.arch()));
+    }
+  } catch (error) {
+    console.error(error);
+  }
 }
 
 function pack(plat, arch, cb) {
   // there is no darwin ia32 electron
   if (plat === 'darwin' && arch === 'ia32') return;
 
-  var opts = assign({}, DEFAULT_OPTS, {
+  const iconObj = {
+    icon: DEFAULT_OPTS.icon + (() => {
+      let extension = '.png';
+      if (plat === 'darwin') {
+        extension = '.icns';
+      } else if (plat === 'win32') {
+        extension = '.ico';
+      }
+      return extension;
+    })()
+  };
+
+  const opts = Object.assign({}, DEFAULT_OPTS, iconObj, {
     platform: plat,
-    arch: arch,
-    out: 'release/' + plat + '-' + arch
+    arch,
+    prune: true,
+    'app-version': pkg.version || DEFAULT_OPTS.version,
+    out: `release/${plat}-${arch}`
   });
 
   packager(opts, cb);
@@ -92,8 +125,8 @@ function pack(plat, arch, cb) {
 
 
 function log(plat, arch) {
-  return function(err, filepath) {
+  return (err, filepath) => {
     if (err) return console.error(err);
-    console.log(plat + '-' + arch + ' finished!');
+    console.log(`${plat}-${arch} finished!`);
   };
 }