+++ server.js
@@ -1,30 +1,50 @@
-var path = require('path');
-var express = require('express');
-var webpack = require('webpack');
-var config = require('./webpack.config.development');
 
-var app = express();
-var compiler = webpack(config);
+/* eslint no-console: 0 */
+import path from 'path';
+import express from 'express';
+import webpack from 'webpack';
+import webpackDevMiddleware from 'webpack-dev-middleware';
+import webpackHotMiddleware from 'webpack-hot-middleware';
 
-app.use(require('webpack-dev-middleware')(compiler, {
-  publicPath: config.output.publicPath,
+import config from './webpack.config.development';
+
+const app = express();
+const compiler = webpack(config);
+const PORT = process.env.PORT || 3000;
+
+const wdm = webpackDevMiddleware(compiler, {
+   publicPath: config.output.publicPath,
   stats: {
     colors: true
   }
-}));
-
-app.use(require('webpack-hot-middleware')(compiler));
-
-
-app.get('*', function(req, res) {
-  res.sendFile(path.join(__dirname, 'app', 'hot-dev-app.html'));
 });
 
-app.listen(3000, 'localhost', function(err) {
+
+
+// app.get('*', function(req, res) {
+//   res.sendFile(path.join(__dirname, 'app', 'hot-dev-app.html'));
+// });
+
+
+
+app.use(wdm);
+
+app.use(webpackHotMiddleware(compiler));
+
+const server = app.listen(PORT, 'localhost', err => {
   if (err) {
-    console.log("errorListen",err);
+    console.error(err);
     return;
   }
 
-  console.log('Listening at http://localhost:3000');
+  console.log(`Listening at http://localhost:${PORT}`);
+});
+
+process.on('SIGTERM', () => {
+  console.log('Stopping dev server');
+  wdm.close();
+  server.close(() => {
+    process.exit(0);
+  });
+
 });