+++ test/e2e.js
@@ -1,129 +1,104 @@
-import webdriver from 'selenium-webdriver';
+
+import { Application } from 'spectron';
 import { expect } from 'chai';
-import electronPath from 'electron-prebuilt';
-import homeStyles from '../app/components/Home.module.css';
-import counterStyles from '../app/components/Counter.module.css';
+import electronPath from 'electron';
+import homeStyles from '../app/components/Home.css';
+import counterStyles from '../app/components/Counter.css';
+
+const delay = time => new Promise(resolve => setTimeout(resolve, time));
 
 describe('main window', function spec() {
-  before((done) => {
-    this.timeout(5000);
-    this.driver = new webdriver.Builder()
-      .usingServer('http://localhost:9515')
-      .withCapabilities({
-        chromeOptions: {
-          binary: electronPath,
-          args: [ 'app=.' ]
-        }
-      })
-      .forBrowser('electron')
-      .build();
-    done();
+  this.timeout(5000);
+
+  before(async () => {
+    this.app = new Application({
+      path: electronPath,
+      args: ['.'],
+    });
+    return this.app.start();
   });
 
-  after((done) => {
-    this.timeout(10000);
-    this.driver.quit().then(done);
+  after(() => {
+    if (this.app && this.app.isRunning()) {
+      return this.app.stop();
+    }
   });
 
-  const findCounter = () => {
-    return this.driver.findElement(webdriver.By.className(counterStyles.counter));
+  const findCounter = () => this.app.client.element(`.${counterStyles.counter}`);
+
+  const findButtons = async () => {
+    const { value } = await this.app.client.elements(`.${counterStyles.btn}`);
+    return value.map(btn => btn.ELEMENT);
   };
 
-  const findButtons = () => {
-    return this.driver.findElements(webdriver.By.className(counterStyles.btn));
-  };
+  it('should open window', async () => {
+    const { client, browserWindow } = this.app;
 
-  it('should open window', (done) => {
-    async () => {
-      const title = await this.driver.getTitle();
-      expect(title).to.equal('Hello Electron React!');
-      done();
-    }().catch(done);
+    await client.waitUntilWindowLoaded();
+    await delay(500);
+    const title = await browserWindow.getTitle();
+    expect(title).to.equal('Hello Electron React!');
   });
 
-  it('should to Counter with click "to Counter" link', (done) => {
-    async () => {
-      const link = await this.driver.findElement(webdriver.By.css(`.${homeStyles.container} > a`));
-      link.click();
+  it('should to Counter with click "to Counter" link', async () => {
+    const { client } = this.app;
 
-      const counter = await findCounter();
-      expect(await counter.getText()).to.equal('0');
-      done();
-    }().catch(done);
+    await client.click(`.${homeStyles.container} > a`);
+    expect(await findCounter().getText()).to.equal('0');
   });
 
-  it('should display updated count after increment button click', (done) => {
-    async () => {
-      const buttons = await findButtons();
-      buttons[0].click();
+  it('should display updated count after increment button click', async () => {
+    const { client } = this.app;
 
-      const counter = await findCounter();
-      expect(await counter.getText()).to.equal('1');
-      done();
-    }().catch(done);
+    const buttons = await findButtons();
+    await client.elementIdClick(buttons[0]);  // +
+    expect(await findCounter().getText()).to.equal('1');
   });
 
-  it('should display updated count after descrement button click', (done) => {
-    async () => {
-      const buttons = await findButtons();
-      const counter = await findCounter();
+  it('should display updated count after descrement button click', async () => {
+    const { client } = this.app;
 
-      buttons[1].click();  // -
-
-      expect(await counter.getText()).to.equal('0');
-      done();
-    }().catch(done);
+    const buttons = await findButtons();
+    await client.elementIdClick(buttons[1]);  // -
+    expect(await findCounter().getText()).to.equal('0');
   });
 
-  it('shouldnt change if even and if odd button clicked', (done) => {
-    async () => {
-      const buttons = await findButtons();
-      const counter = await findCounter();
-      buttons[2].click();  // odd
+  it('shouldnt change if even and if odd button clicked', async () => {
+    const { client } = this.app;
 
-      expect(await counter.getText()).to.equal('0');
-
-      done();
-    }().catch(done);
+    const buttons = await findButtons();
+    await client.elementIdClick(buttons[2]);  // odd
+    expect(await findCounter().getText()).to.equal('0');
   });
 
-  it('should change if odd and if odd button clicked', (done) => {
-    async () => {
-      const buttons = await findButtons();
-      const counter = await findCounter();
+  it('should change if odd and if odd button clicked', async () => {
+    const { client } = this.app;
 
-      buttons[0].click();  // +
-      buttons[2].click();  // odd
-
-      expect(await counter.getText()).to.equal('2');
-
-      done();
-    }().catch(done);
+    const buttons = await findButtons();
+    await client.elementIdClick(buttons[0]);  // +
+    await client.elementIdClick(buttons[2]);  // odd
+    expect(await findCounter().getText()).to.equal('2');
   });
 
-  it('should change if async button clicked and a second later', (done) => {
-    async () => {
-      const buttons = await findButtons();
-      const counter = await findCounter();
-      buttons[3].click();  // async
+  it('should change if async button clicked and a second later', async () => {
+    const { client } = this.app;
 
-      expect(await counter.getText()).to.equal('2');
-
-      await this.driver.wait(() =>
-        counter.getText().then(text => text === '3' )
-      , 1000, 'count not as expected');
-
-      done();
-    }().catch(done);
+    const buttons = await findButtons();
+    await client.elementIdClick(buttons[3]);  // async
+    expect(await findCounter().getText()).to.equal('2');
+    await delay(1000);
+    expect(await findCounter().getText()).to.equal('3');
   });
 
-  it('should back to home if back button clicked', (done) => {
-    async () => {
-      const link = await this.driver.findElement(webdriver.By.css(`.${counterStyles.backButton} > a`));
-      link.click();
+  it('should back to home if back button clicked', async () => {
+    const { client } = this.app;
+    await client.element(
+      `.${counterStyles.backButton} > a`
+    ).click();
 
-      await this.driver.findElement(webdriver.By.className(homeStyles.container));
-      done();
-    }().catch(done);
+    expect(
+      await client.isExisting(`.${homeStyles.container}`)
+    ).to.be.true;
+
   });
 });