+++ test/setup.js
@@ -1,10 +1,21 @@
-import { jsdom } from 'jsdom';
-import hook from 'css-modules-require-hook';
 
-hook({
-  generateScopedName: '[name]__[local]___[hash:base64:5]'
-});
+import 'babel-polyfill';
+import { jsdom } from 'jsdom';
+
 
 global.document = jsdom('<!doctype html><html><body></body></html>');
 global.window = document.defaultView;
 global.navigator = global.window.navigator;
+
+window.localStorage = window.sessionStorage = {
+  getItem(key) {
+    return this[key];
+  },
+  setItem(key, value) {
+    this[key] = value;
+  },
+  removeItem(key) {
+    this[key] = undefined;
+  },
+};
+>>>>>>> 9d0782912ae4100825089e1678a684966bedee59