+++ webpack.config.base.js
@@ -1,29 +1,32 @@
-var path = require('path');
 
+import path from 'path';
 
-
-var config = {
+export default {
   module: {
     loaders: [{
       test: /\.jsx?$/,
-      loaders: ['react-hot','babel-loader'],
+      loaders: [/*'react-hot',*/'babel-loader'],
       exclude: /node_modules/
-    }]   
+    }, {
+      test: /\.json$/,
+      loader: 'json-loader'
+    }]
   },
   output: {
     path: path.join(__dirname, 'dist'),
-    filename: "[name]bundle.js",
+    filename: 'bundle.js',
     libraryTarget: 'commonjs2',
-    publicPath: 'http://localhost:3000/dist/'
+      //  publicPath: 'http://localhost:3000/dist/'
   },
   resolve: {
-    extensions: ['', '.js', '.jsx'],
+    extensions: ['', '.js', '.jsx', '.json'],
     packageMains: ['webpack', 'browser', 'web', 'browserify', ['jam', 'main'], 'main']
   },
   plugins: [
 
   ],
   externals: [
+
     // put your node 3rd party libraries which can't be built with webpack here (mysql, mongodb, and so on..)
     // "node-osc",
     "pouchdb",
@@ -35,6 +38,3 @@ var config = {
   ]
 };
 
-// config.output.;
-
-module.exports = config;